chore: adopt ruff 0.16's rule set, and declare it so it stops moving
CI installs ruff unpinned and 0.16 expanded its default rule set from a handful
to 413 rules. Main went from green to 1,678 errors overnight without a line of
our code changing. Verified: a clean worktree of origin/main, checked with
0.16.3, reports 1,678 — this is not one branch's problem.

The repo had NO ruff configuration at all, which is the actual root cause: every
release got to redefine what "lint clean" means here. pyproject.toml now carries
a [tool.ruff.lint] section, and every declined rule has a written reason.

  1,678 -> 1,239 auto-fixed
        ->   142 after declining the rules that fight deliberate house style
        ->     0

Declined, with reasons in the config: C408 (292) — `dict(size=12)` is Plotly's
own idiom and plots.py alone has 220 nested in figure specs; BLE001 (31) —
broad excepts that are deliberate and documented at each site; ISC004 (21) —
adjacent strings in a list is the shape of a missing comma, so all 21 were read
individually and every one is an intentional line wrap; plus TRY004, S110/S112,
DTZ005/006, EXE001, SIM115.

TWO REGRESSIONS THE AUTO-FIXER INTRODUCED, both caught by tests:

  * SIM401 and RUF019 rewrite `"k" in x and x["k"]` to `x.get("k")`. Correct for
    st.session_state, WRONG for AppTest.session_state, whose proxy has no .get
    and raises AttributeError. Broke 5 tests. Both are ignored for tests/ —
    together, because silencing one leaves the other to make the same edit.

  * `--fix` deleted 21 `# noqa` directives as "unused", including the ones
    guarding api.py's deliberately load-bearing import order. They were only
    unused because 0.16 does not enable E402/E712/N802 by default; under any
    older ruff the code was then broken. All 21 restored, and those three rules
    are now named in extend-select so the suppressions stay live either way.

Verified clean on ruff 0.14.13, 0.15.14 AND 0.16.3 — version skew was the whole
problem, so one version proves nothing. Full suite 2059 passed, 0 failed
(2052+7, 2054+5, 2059+0 across three runs: no tests lost).

Note for review: 0.16 also formats Python blocks inside Markdown, which older
versions did not. README.md and four docs pages are reformatted as a result, and
one README one-liner now wraps across three lines. Cosmetic and revertible, but
it is a documentation change inside a lint commit.

+++ paper/paper_figs.py
@@ -97,7 +97,7 @@ def replay_fig(words, fixations):
     t0 = tf["timestamp_ms"].iloc[0]
     frames = []
     for frac in (0.33, 0.66, 1.0):
-        k = max(2, int(round(n * frac)))
+        k = max(2, round(n * frac))
         keep = set(tf["order_in_trial"].iloc[:k])
         sub = fixations[
             ~((fixations.participant_id == pid) & (fixations.trial_id == tid))
